Move element alignment to its own file
